requirements.txt
pypdf2
sentence-transformers
faiss-cpu
streamlit
langchain
langchain-community
numpy
torch
groq
python-dotenv
